Disable syntax highlighting for rush publish usage
In GitHub-flavored markdown, when no language is specified for a code block, the renderer will attempt to determine the language from the content. In cases where a code block contains non-language content (i.e., the 'usage' message for a command), this leads to highlighting of seemingly random words in seemingly random colors throughout the document.

To suppress language auto-detection for a code block, specify an unrecognized language. See the [GitHub-flavored Markdown spec](https://github.github.com/gfm/#info-string) for examples.

This PR specifies the language as `unknown`. The same issue probably occurs in other documents and should be fixed in those documents as well.